Split Matter logic from MDNS and add caching
